+++ packages/calypso-build/jest-preset.js
@@ -5,7 +5,7 @@ const path = require( 'path' );
 const { defaults } = require( 'jest-config' );
 
 module.exports = {
-	setupFilesAfterEnv: [ path.join( __dirname, 'jest', 'setup.js' ) ],
+	setupFilesAfterEnv: [ path.join( __dirname, 'jest', 'setup.js' ), 'jest-enzyme' ],
 	snapshotSerializers: [ 'enzyme-to-json/serializer' ],
 	testEnvironment: 'node',
 	testMatch: [ '<rootDir>/**/test/*.[jt]s?(x)', '!**/.eslintrc.*' ],