+++ packages/calypso-ui/src/button/test/index.js
@@ -1,49 +1,45 @@
-/** @format */
 /**
  * External dependencies
  */
-import { expect } from 'chai';
-import { shallow } from 'enzyme';
-import Gridicon from 'components/gridicon';
 import React from 'react';
+import Gridicon from 'gridicons'; // eslint-disable-line no-restricted-imports
+import { shallow } from 'enzyme';
 
 /**
  * Internal dependencies
  */
-import Button from '../';
+import Button from '..';
 
 describe( 'Button', () => {
 	describe( 'renders', () => {
 		test( 'with modifiers', () => {
 			const button = shallow( <Button scary primary borderless compact /> );
-			expect( button ).to.have.className( 'is-compact' );
-			expect( button ).to.have.className( 'is-primary' );
-			expect( button ).to.have.className( 'is-scary' );
-			expect( button ).to.have.className( 'is-borderless' );
+			expect( button ).toHaveClassName( 'is-compact' );
+			expect( button ).toHaveClassName( 'is-primary' );
+			expect( button ).toHaveClassName( 'is-scary' );
+			expect( button ).toHaveClassName( 'is-borderless' );
 		} );
 
 		test( 'without modifiers', () => {
 			const button = shallow( <Button /> );
-			expect( button ).to.have.className( 'button' );
-			expect( button ).to.not.have.className( 'is-compact' );
-			expect( button ).to.not.have.className( 'is-primary' );
-			expect( button ).to.not.have.className( 'is-scary' );
-			expect( button ).to.not.have.className( 'is-borderless' );
+			expect( button ).toHaveClassName( 'button' );
+			expect( button ).not.toHaveClassName( 'is-compact' );
+			expect( button ).not.toHaveClassName( 'is-primary' );
+			expect( button ).not.toHaveClassName( 'is-scary' );
+			expect( button ).not.toHaveClassName( 'is-borderless' );
 		} );
 
 		test( 'disabled', () => {
 			const button = shallow( <Button disabled /> );
-			expect( button ).to.be.disabled;
+			expect( button ).toBeDisabled();
 		} );
 
 		test( 'with child', () => {
 			const iconType = 'arrow-left';
 			const icon = <Gridicon size={ 18 } icon={ iconType } />;
 			const button = shallow( <Button>{ icon }</Button> );
-			expect( button ).to.contain( icon );
-			expect( button.find( Gridicon ) )
-				.to.have.prop( 'icon' )
-				.equal( iconType );
+			expect( button ).toContainReact( icon );
+			expect( button.find( Gridicon ) ).toHaveProp( 'icon', iconType );
 		} );
 	} );
 
@@ -51,16 +47,14 @@ describe( 'Button', () => {
 		test( 'renders as a link', () => {
 			const button = shallow( <Button href="https://wordpress.com/" /> );
 
-			expect( button ).to.match( 'a' );
-			expect( button )
-				.to.have.prop( 'href' )
-				.equal( 'https://wordpress.com/' );
+			expect( button ).toMatchSelector( 'a' );
+			expect( button ).toHaveProp( 'href', 'https://wordpress.com/' );
 		} );
 
 		test( 'ignores type prop and renders a link without type attribute', () => {
 			const button = shallow( <Button href="https://wordpress.com/" type="submit" /> );
 
-			expect( button ).to.not.have.prop( 'type' );
+			expect( button ).not.toHaveProp( 'type' );
 		} );
 
 		test( 'including target and rel props renders a link with target and rel attributes', () => {
@@ -68,29 +62,17 @@ describe( 'Button', () => {
 				<Button href="https://wordpress.com/" target="_blank" rel="noopener noreferrer" />
 			);
 
-			expect( button )
-				.to.have.prop( 'target' )
-				.equal( '_blank' );
-			expect( button )
-				.to.have.prop( 'rel' )
-				.match( /\bnoopener\b/ );
-			expect( button )
-				.to.have.prop( 'rel' )
-				.match( /\bnoreferrer\b/ );
+			expect( button ).toHaveProp( 'target', '_blank' );
+			expect( button ).toHaveProp( 'rel', expect.stringMatching( /\bnoopener\b/ ) );
+			expect( button ).toHaveProp( 'rel', expect.stringMatching( /\bnoreferrer\b/ ) );
 		} );
 
 		test( 'adds noopener noreferrer rel if target is specified', () => {
 			const button = shallow( <Button href="https://wordpress.com/" target="_blank" /> );
 
-			expect( button )
-				.to.have.prop( 'target' )
-				.equal( '_blank' );
-			expect( button )
-				.to.have.prop( 'rel' )
-				.match( /\bnoopener\b/ );
-			expect( button )
-				.to.have.prop( 'rel' )
-				.match( /\bnoreferrer\b/ );
+			expect( button ).toHaveProp( 'target', '_blank' );
+			expect( button ).toHaveProp( 'rel', expect.stringMatching( /\bnoopener\b/ ) );
+			expect( button ).toHaveProp( 'rel', expect.stringMatching( /\bnoreferrer\b/ ) );
 		} );
 	} );
 
@@ -98,14 +80,12 @@ describe( 'Button', () => {
 		const button = shallow( <Button target="_blank" rel="noopener noreferrer" /> );
 
 		test( 'renders as a button', () => {
-			expect( button ).to.match( 'button' );
-			expect( button ).to.not.have.prop( 'href' );
+			expect( button ).toMatchSelector( 'button' );
+			expect( button ).not.toHaveProp( 'href' );
 		} );
 
 		test( 'renders button with type attribute set to "button" by default', () => {
-			expect( button )
-				.to.have.prop( 'type' )
-				.equal( 'button' );
+			expect( button ).toHaveProp( 'type', 'button' );
 		} );
 
 		test( 'renders button with type attribute set to type prop if specified', () => {
@@ -114,14 +94,12 @@ describe( 'Button', () => {
 				<Button target="_blank" rel="noopener noreferrer" type={ typeProp } />
 			);
 
-			expect( submitButton )
-				.to.have.prop( 'type' )
-				.equal( typeProp );
+			expect( submitButton ).toHaveProp( 'type', typeProp );
 		} );
 
 		test( 'renders button without rel and target attributes', () => {
-			expect( button ).to.not.have.prop( 'target' );
-			expect( button ).to.not.have.prop( 'rel' );
+			expect( button ).not.toHaveProp( 'target' );
+			expect( button ).not.toHaveProp( 'rel' );
 		} );
 	} );
 } );